+++ tokio-postgres/src/row.rs
@@ -79,9 +79,9 @@ impl RowIndex for str {
     }
 }
 
-impl<'a, T> Sealed for &'a T where T: ?Sized + Sealed {}
+impl<T> Sealed for &T where T: ?Sized + Sealed {}
 
-impl<'a, T> RowIndex for &'a T
+impl<T> RowIndex for &T
 where
     T: ?Sized + RowIndex,
 {
@@ -95,6 +95,7 @@ where
 }
 
 /// A row of data returned from the database by a query.
+#[derive(Clone)]
 pub struct Row {
     statement: Statement,
     body: DataRowBody,