+++ postgres-protocol/CHANGELOG.md
@@ -1,6 +1,27 @@
 # Change Log
 
-## v0.6.6 -2023-08-19
+## v0.6.8 - 2025-02-02
+
+### Changed
+
+* Upgraded `getrandom`.
+
+## v0.6.7 - 2024-07-21
+
+### Deprecated
+
+* Deprecated `ErrorField::value`.
+
+### Added
+
+* Added a `Clone` implementation for `DataRowBody`.
+* Added `ErrorField::value_bytes`.
+
+### Changed
+
+* Upgraded `base64`.
+
+## v0.6.6 - 2023-08-19
 
 ### Added