+++ postgres-types/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Change Log
 
+## Unreleased
+
+## v0.2.9 - 2025-02-02
+
+### Added
+
+* Added support for `cidr` 0.3 via the `with-cidr-0_3` feature.
+
+### Fixed
+
+* Fixed deserialization of out of bounds inputs to `time` 0.3 types to return an error rather than panic.
+
+## v0.2.8 - 2024-09-15
+
+### Added
+
+* Added support for `jiff` 0.1 via the `with-jiff-01` feature.
+
+## v0.2.7 - 2024-07-21
+
+### Added
+
+* Added `Default` implementation for `Json`.
+* Added a `js` feature for WASM compatibility.
+
+### Changed
+
+* `FromStr` implementation for `PgLsn` no longer allocates a `Vec` when splitting an lsn string on it's `/`.
+* The `eui48-1` feature no longer enables default features of the `eui48` library.
+
 ## v0.2.6 - 2023-08-19
 
 ### Fixed